requirements.txt
pypdf>=3.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.0
transformers>=4.30.0
torch>=2.0.0
streamlit>=1.25.0
notebook>=7.0.0
ipykernel>=6.0.0
